Fixed conversion warning
- Changed InitRadius variable from the DungeonGenerator class from float to int
- Changed all getters and setters that use this InitRadius variable from float to int

+++ Project/Dungeon.cpp
@@ -17,7 +17,7 @@ void Dungeon::SetSeed(int seed)
 	m_Generator.SetSeed(seed);
 }
 
-void Dungeon::SetInitialRadius(float radius)
+void Dungeon::SetInitialRadius(int radius)
 {
 	m_Generator.SetInitRadius(radius);
 }
